fix: move domain and sso type to ee
